Move _urls_for_js to utils.py (circular imports)
