+++ umap/views.py
@@ -390,22 +390,6 @@ ajax_proxy = AjaxProxy.as_view()
 # ############## #
 
 
-def _urls_for_js(urls=None):
-    """
-    Return templated URLs prepared for javascript.
-    """
-    if urls is None:
-        # prevent circular import
-        from .urls import i18n_urls, urlpatterns
-
-        urls = [
-            url.name for url in urlpatterns + i18n_urls if getattr(url, "name", None)
-        ]
-    urls = dict(zip(urls, [get_uri_template(url) for url in urls]))
-    urls.update(getattr(settings, "UMAP_EXTRA_URLS", {}))
-    return urls
-
-
 def simple_json_response(**kwargs):
     return HttpResponse(json.dumps(kwargs), content_type="application/json")
